requirements.txt
torch
torchvision
timm
numpy
pandas
polars
opencv-python
pydicom
scikit-learn
tqdm
kaggle
python-dotenv
